Issue 13575: Add support for mp.jwt.decrypt.key.location in MP JWT 1.2 - test code
