+++ dev/com.ibm.ws.security.common.jsonwebkey/test/com/ibm/ws/security/common/jwk/impl/JWKSetTest.java
@@ -10,10 +10,12 @@
  *******************************************************************************/
 package com.ibm.ws.security.common.jwk.impl;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 
 import java.net.URI;
 import java.net.URL;
+import java.security.Key;
 import java.security.PrivateKey;
 import java.security.PublicKey;
 import java.util.ArrayList;
@@ -27,6 +29,7 @@ import org.junit.Before;
 import org.junit.Test;
 
 import com.ibm.json.java.JSONObject;
+import com.ibm.ws.security.common.jwk.impl.JwKRetriever.JwkKeyType;
 import com.ibm.ws.security.common.jwk.interfaces.JWK;
 
 public class JWKSetTest {
@@ -81,9 +84,10 @@ public class JWKSetTest {
         setId = uri.toString();
         jwkSet.add(setId, jwk);
 
-        PublicKey publicKey = jwkSet.getPublicKeyBySetId(setId);
+        Key publicKey = jwkSet.getKeyBySetId(setId, JwkKeyType.PUBLIC);
 
         assertNotNull("There must a public key.", publicKey);
+        assertTrue("Returned key was not a PublicKey: " + publicKey, publicKey instanceof PublicKey);
     }
 
     @Test
@@ -91,42 +95,44 @@ public class JWKSetTest {
         setId = uri.toString();
         jwkSet.add(setId, jwk);
 
-        PublicKey publicKey = jwkSet.getPublicKeyBySetIdAndKid(setId, kid);
+        Key publicKey = jwkSet.getKeyBySetIdAndKid(setId, kid, JwkKeyType.PUBLIC);
 
         assertNotNull("There must a public key.", publicKey);
+        assertTrue("Returned key was not a PublicKey: " + publicKey, publicKey instanceof PublicKey);
     }
-
 
     @Test
     public void testGetPublicKeyByUriAndx5t() throws Exception {
         setId = uri.toString();
         jwkSet.add(setId, jwk);
 
-        PublicKey publicKey = jwkSet.getPublicKeyBySetIdAndx5t(setId, x5t);
+        Key publicKey = jwkSet.getKeyBySetIdAndx5t(setId, x5t, JwkKeyType.PUBLIC);
 
         assertNotNull("There must a public key.", publicKey);
+        assertTrue("Returned key was not a PublicKey: " + publicKey, publicKey instanceof PublicKey);
     }
-    
+
     @Test
     public void testRemoveStaleEntries() throws Exception {
         ArrayList<JWK> al = new ArrayList<JWK>();
         al.add(new TestJWK());
         TestJWK j2 = new TestJWK();
-        j2.created =  System.currentTimeMillis();
+        j2.created = System.currentTimeMillis();
         al.add(j2);
         al.add(new TestJWK());
         j2 = new TestJWK();
-        j2.created =  System.currentTimeMillis();
+        j2.created = System.currentTimeMillis();
         al.add(j2);
-        
+
         JWKSet testSet = new JWKSet();
         testSet.removeStaleEntries(al);
         assertTrue("Expected two entries to be removed", al.size() == 2);
-        
+
     }
-    
-    class TestJWK implements JWK{
+
+    class TestJWK implements JWK {
         long created = 0l;
+
         @Override
         public String getKeyID() {
             return null;
@@ -143,7 +149,7 @@ public class JWKSetTest {
         }
 
         @Override
-        public String getKeyUse() {            
+        public String getKeyUse() {
             return null;
         }
