+++ dev/io.openliberty.microprofile.jwt.1.2.internal_fat_tck/publish/servers/tckAudEnv/server.env
@@ -14,5 +14,3 @@
 
 #mp.jwt.verify.publickey.location=publicKey.pem
 #mp.jwt.verify.issuer=https://server.example.com
-mp_jwt_verify_publickey_location=publicKey.pem
-mp_jwt_verify_issuer=https://server.example.com
